add authentication for /routes endpoints

+++ app/routers/route.py
@@ -5,8 +5,9 @@ from sqlmodel import select
 
 import datetime as dt
 
-from app.schemas import RouteBase, Route, RoutePublic, RouteCreate
+from app.schemas import RouteBase, Route, RoutePublic, RouteCreate, Role
 from app.services.database import SessionDependency
+from app.services import auth
 
 router = APIRouter()
 
@@ -22,8 +23,9 @@ def get_route(session: SessionDependency, route_id: int) -> RoutePublic:
     return route
 
 @router.post("/routes/create", response_model=RoutePublic, status_code=status.HTTP_201_CREATED)
-def create_route(session: SessionDependency, route_create: RouteCreate) -> RoutePublic:
-    # add auth guards
+def create_route(current_user: auth.CurrentUser, session: SessionDependency, route_create: RouteCreate) -> RoutePublic:
+    if current_user.role != Role.ADMIN:
+        raise HTTPException(status_code=403, detail="Admin role required")
 
     route_create = RouteCreate.model_validate(route_create)
     route = Route(
@@ -35,14 +37,16 @@ def create_route(session: SessionDependency, route_create: RouteCreate) -> Route
     return route
 
 @router.get("/routes/suggest", response_model=list[RoutePublic])
-def suggest_routes(session: SessionDependency, offset: int = 0, limit: Annotated[int, Query(le=100)] = 100) -> list[RoutePublic]:
-    #add auth guards
+def suggest_routes(current_user: auth.CurrentUser, session: SessionDependency, offset: int = 0, limit: Annotated[int, Query(le=100)] = 100) -> list[RoutePublic]:
+    if current_user.role != Role.ADMIN:
+            raise HTTPException(status_code=403, detail="Admin role required")
     routes = session.exec(select(Route).offset(offset).limit(limit).where(Route.is_approved == False))
     return routes
 
 @router.get("/routes/approve/{route_id}", response_model=RoutePublic)
-def approve_route(session: SessionDependency, route_id: int) -> RoutePublic:
-    #add auth guards
+def approve_route(current_user: auth.CurrentUser, session: SessionDependency, route_id: int) -> RoutePublic:
+    if current_user.role != Role.ADMIN:
+            raise HTTPException(status_code=403, detail="Admin role required")
     route = session.get(Route, route_id)
     if not route: raise HTTPException(status_code=404, detail="Route not found")
     route_data = route.model_dump()
@@ -55,8 +59,9 @@ def approve_route(session: SessionDependency, route_id: int) -> RoutePublic:
     return route
 
 @router.get("/routes/delete/{route_id}", response_model=RoutePublic)
-def delete_route(session: SessionDependency, route_id: int) -> RoutePublic:
-    #add auth guards
+def delete_route(current_user: auth.CurrentUser, session: SessionDependency, route_id: int) -> RoutePublic:
+    if current_user.role != Role.ADMIN:
+            raise HTTPException(status_code=403, detail="Admin role required")
     route = session.get(Route, route_id)
     if not route: raise HTTPException(status_code=404, detail="Route not found")
     session.delete(route)
